9 of 12? subworkflow submodules added
Broad workflow needs to wrangled from wdl to cwl so that I can integrate it better.
SMuFIN seems like it was included in pcawg at last minute and I will have to build up a wrapper myself to include it.
Broad PoN Filter is meant to be called on a population of calls. I still haven't decided if I'm going to try to scatter a whole population of inputs.